Address review polish on sync-fsrs-and-meanings
- Debounce the FSRS-settings push in App.tsx (500ms) so a slider drag
  collapses into one outbox op instead of dozens. Was the only real
  efficiency hazard flagged in review.
- Hoist FSRSSettings interface from the store into db/schema and have
  Deck.fsrsSettings be Partial<FSRSSettings> instead of the loose
  Record<string, unknown>. Drops the as-unknown-as cast at the App.tsx
  call site and gives compile-time keys for the FIELD_MAP whitelist.
  Re-exports FSRSSettings from the store for callers that already
  import it from there.
- Replace the (merged as any)[k] / (blob as any)[k] casts in
  hydrateFSRSSettingsFromBlob with a single Partial<FSRSSettings> view.
- Drop redundant {...DEFAULTS} as FSRSSettings cast.
- Add multi-deck TODO on the syncEngine hydrate loop so the global-
  store-vs-per-deck-settings tension is visible once decks branch.
