+++ src/stores/fsrsSettingsStore.ts
@@ -1,16 +1,10 @@
 import { create } from 'zustand';
 import type { Steps } from 'ts-fsrs';
+import type { FSRSSettings } from '../db/schema';
+
+export type { FSRSSettings };
 
 const STORAGE_KEY = 'mandao_fsrs_settings';
-
-export interface FSRSSettings {
-  requestRetention: number;
-  maximumInterval: number;
-  enableFuzz: boolean;
-  enableShortTerm: boolean;
-  learningSteps: string[];
-  relearningSteps: string[];
-}
 
 const DEFAULTS: FSRSSettings = {
   requestRetention: 0.9,
@@ -58,9 +52,9 @@ export const useFSRSSettingsStore = create<FSRSSettingsState>((set, get) => ({
 /** Pull only the FSRSSettings shape out of the store (drop action methods). */
 export function getFSRSSettings(): FSRSSettings {
   const state = useFSRSSettingsStore.getState();
-  const out: Record<string, unknown> = {};
-  for (const k of SETTING_KEYS) out[k] = state[k];
-  return out as unknown as FSRSSettings;
+  const out = {} as FSRSSettings;
+  for (const k of SETTING_KEYS) (out[k] as FSRSSettings[typeof k]) = state[k];
+  return out;
 }
 
 /**
@@ -74,15 +68,13 @@ export function isHydratingFSRSSettings(): boolean {
 }
 export function hydrateFSRSSettingsFromBlob(blob: Record<string, unknown> | undefined): void {
   if (!blob || Object.keys(blob).length === 0) return;
-  const merged = { ...DEFAULTS } as FSRSSettings;
-  let changed = false;
+  const merged: FSRSSettings = { ...DEFAULTS };
+  const partial = blob as Partial<FSRSSettings>;
   for (const k of SETTING_KEYS) {
-    if (k in blob) {
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      (merged as any)[k] = (blob as any)[k];
-    }
+    if (k in partial) (merged[k] as FSRSSettings[typeof k]) = partial[k] as FSRSSettings[typeof k];
   }
   const cur = getFSRSSettings();
+  let changed = false;
   for (const k of SETTING_KEYS) {
     if (JSON.stringify(cur[k]) !== JSON.stringify(merged[k])) { changed = true; break; }
   }
